tools/requirements.txt
chardet==5.2.0
pandas==2.1.4
psycopg2-binary==2.9.9 